+++ src/main/java/com/sanbosillok/sanbosillokserver/api/post/dto/PostTitleResponse.java
@@ -1,6 +1,6 @@
-package com.sanbosillok.sanbosillokserver.post.dto;
+package com.sanbosillok.sanbosillokserver.api.post.dto;
 
-import com.sanbosillok.sanbosillokserver.post.domain.Post;
+import com.sanbosillok.sanbosillokserver.api.post.domain.Post;
 import lombok.Data;
 import lombok.NoArgsConstructor;
